Fix issue regarding metadata reading
Needed to use an io.Reader to keep metadata from being changed to a
base64 encoded string
